+++ proofs/examples/two-chip.rs
@@ -260,7 +260,7 @@ impl<F: Field> MulChip<F> {
         for column in &advice {
             meta.enable_equality(*column);
         }
-        let s_mul = meta.selector();
+        let s_mul = meta.complex_selector();
 
         // Define our multiplication gate!
         meta.create_gate("mul", |meta| {
@@ -285,7 +285,7 @@ impl<F: Field> MulChip<F> {
             // has the following properties:
             // - When s_mul = 0, any value is allowed in lhs, rhs, and out.
             // - When s_mul != 0, this constrains lhs * rhs = out.
-            Constraints::with_selector(s_mul, vec![lhs * rhs - out])
+            Constraints::with_additive_selector(s_mul, vec![lhs * rhs - out])
         });
 
         MulConfig { advice, s_mul }