+++ proofs/src/plonk/circuit.rs
@@ -13,10 +13,11 @@ use std::{
 use ff::Field;
 use sealed::SealedPhase;
 
-use super::{lookup, permutation, Error};
+use super::{lookup, permutation, trash, Error};
 use crate::{
     circuit::{layouter::SyncDeps, Layouter, Region, Value},
     dev::metadata,
+    plonk::trash::Argument,
     poly::Rotation,
     utils::rational::Rational,
 };
@@ -1424,9 +1425,10 @@ impl<F: Field> From<Expression<F>> for Vec<Constraint<F>> {
     }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Debug)]
 enum SelectorType {
     Multiplicative(Selector),
+    Additive(Selector),
     None,
 }
 
@@ -1478,6 +1480,23 @@ impl<F: Field> Constraints<F> {
     pub fn with_selector<I: Into<Constraint<F>>>(selector: Selector, constraints: Vec<I>) -> Self {
         Constraints {
             selector: SelectorType::Multiplicative(selector),
+            constraints: constraints.into_iter().map(|c| c.into()).collect(),
+        }
+    }
+
+    /// Constructs a set of constraints that are controlled by the given
+    /// selector. The selector is involved additively (via a trash::argument)
+    /// and thus, it does not increase the constraints degree.
+    ///
+    /// The enforced constraint will have the form
+    /// `(sum_i r^i constraints[i]) - (1 - selector) * trash`, where `r` is
+    /// a Fiat-Shamir challenge and `trash` is an unrestricted column.
+    pub fn with_additive_selector<I: Into<Constraint<F>>>(
+        selector: Selector,
+        constraints: Vec<I>,
+    ) -> Self {
+        Constraints {
+            selector: SelectorType::Additive(selector),
             constraints: constraints.into_iter().map(|c| c.into()).collect(),
         }
     }
@@ -1565,6 +1584,12 @@ pub struct ConstraintSystem<F: Field> {
     // input expressions and a sequence of table expressions involved in the lookup.
     pub(crate) lookups: Vec<lookup::Argument<F>>,
 
+    // Vector of trash arguments. Each contains a selector and a sequence of expressions
+    // that will all be enforced to be zero when the selector is enabled. This is done
+    // via an argument that involves the selector additively (instead of multiplicatively)
+    // with the help of an extra so-called trash column.
+    pub(crate) trashcans: Vec<trash::Argument<F>>,
+
     // List of indexes of Fixed columns which are associated to a circuit-general Column tied to
     // their annotation.
     pub(crate) general_column_annotations: HashMap<metadata::Column, String>,
@@ -1592,6 +1617,7 @@ pub struct PinnedConstraintSystem<'a, F: Field> {
     fixed_queries: &'a Vec<(Column<Fixed>, Rotation)>,
     permutation: &'a permutation::Argument,
     lookups: &'a Vec<lookup::Argument<F>>,
+    trashcans: &'a Vec<trash::Argument<F>>,
     constants: &'a Vec<Column<Fixed>>,
     minimum_degree: &'a Option<usize>,
 }
@@ -1612,6 +1638,7 @@ impl<F: Field> std::fmt::Debug for PinnedConstraintSystem<'_, F> {
             fixed_queries,
             permutation,
             lookups,
+            trashcans,
             constants,
             minimum_degree,
         } = &self;
@@ -1634,7 +1661,8 @@ impl<F: Field> std::fmt::Debug for PinnedConstraintSystem<'_, F> {
             .field("instance_queries", instance_queries)
             .field("fixed_queries", fixed_queries)
             .field("permutation", permutation)
-            .field("lookups", lookups);
+            .field("lookups", lookups)
+            .field("trashcans", trashcans);
         debug_struct
             .field("constants", constants)
             .field("minimum_degree", minimum_degree);
@@ -1670,6 +1698,7 @@ impl<F: Field> Default for ConstraintSystem<F> {
             instance_queries: Vec::new(),
             permutation: permutation::Argument::new(),
             lookups: Vec::new(),
+            trashcans: Vec::new(),
             general_column_annotations: HashMap::new(),
             constants: vec![],
             minimum_degree: None,
@@ -1696,6 +1725,7 @@ impl<F: Field> ConstraintSystem<F> {
             instance_queries: &self.instance_queries,
             permutation: &self.permutation,
             lookups: &self.lookups,
+            trashcans: &self.trashcans,
             constants: &self.constants,
             minimum_degree: &self.minimum_degree,
         }
@@ -1904,6 +1934,12 @@ impl<F: Field> ConstraintSystem<F> {
     ) {
         let mut cells = VirtualCells::new(self);
         let constraints = constraints(&mut cells);
+
+        assert!(
+            !constraints.constraints.is_empty(),
+            "Gates must contain at least one constraint."
+        );
+
         let (constraint_names, polys): (_, Vec<_>) = cells
             .apply_selector_to_constraints(constraints)
             .into_iter()
@@ -1915,11 +1951,6 @@ impl<F: Field> ConstraintSystem<F> {
 
         let queried_selectors = cells.queried_selectors;
         let queried_cells = cells.queried_cells;
-
-        assert!(
-            !polys.is_empty(),
-            "Gates must contain at least one constraint."
-        );
 
         self.gates.push(Gate {
             name: name.into(),
@@ -1975,8 +2006,7 @@ impl<F: Field> ConstraintSystem<F> {
                 &|selector| {
                     if must_be_nonsimple {
                         // Simple selectors are prohibited from appearing in
-                        // expressions in the lookup argument by
-                        // `ConstraintSystem`.
+                        // expressions in the lookup or trash arguments.
                         assert!(!selector.is_simple());
                     }
 
@@ -1993,13 +2023,13 @@ impl<F: Field> ConstraintSystem<F> {
             );
         }
 
-        // Substitute selectors for the real fixed columns in all gates
+        // Substitute selectors for the real fixed columns in all gates.
         for expr in self.gates.iter_mut().flat_map(|gate| gate.polys.iter_mut()) {
             replace_selectors(expr, selector_replacements, false);
         }
 
         // Substitute non-simple selectors for the real fixed columns in all
-        // lookup expressions
+        // lookup expressions.
         for expr in self.lookups.iter_mut().flat_map(|lookup| {
             lookup
                 .input_expressions
@@ -2007,6 +2037,14 @@ impl<F: Field> ConstraintSystem<F> {
                 .chain(lookup.table_expressions.iter_mut())
         }) {
             replace_selectors(expr, selector_replacements, true);
+        }
+
+        // Substitute selectors in all trash arguments.
+        for trash in self.trashcans.iter_mut() {
+            replace_selectors(&mut trash.selector, selector_replacements, true);
+            for expr in trash.constraint_expressions.iter_mut() {
+                replace_selectors(expr, selector_replacements, true);
+            }
         }
     }
 
@@ -2194,33 +2232,20 @@ impl<F: Field> ConstraintSystem<F> {
     /// Compute the degree of the constraint system (the maximum degree of all
     /// constraints).
     pub fn degree(&self) -> usize {
-        // The permutation argument will serve alongside the gates, so must be
-        // accounted for.
-        let mut degree = self.permutation.required_degree();
-
-        // The lookup argument also serves alongside the gates and must be accounted
-        // for.
-        degree = std::cmp::max(
-            degree,
-            self.lookups
-                .iter()
-                .map(|l| l.required_degree())
-                .max()
-                .unwrap_or(1),
-        );
-
-        // Account for each gate to ensure our quotient polynomial is the
-        // correct degree and that our extended domain is the right size.
-        degree = std::cmp::max(
-            degree,
+        [
+            Some(self.permutation.required_degree()),
+            self.lookups.iter().map(|l| l.required_degree()).max(),
+            self.trashcans.iter().map(|l| l.required_degree()).max(),
             self.gates
                 .iter()
                 .flat_map(|gate| gate.polynomials().iter().map(|poly| poly.degree()))
-                .max()
-                .unwrap_or(0),
-        );
-
-        std::cmp::max(degree, self.minimum_degree.unwrap_or(1))
+                .max(),
+            self.minimum_degree,
+        ]
+        .iter()
+        .filter_map(|&d| d)
+        .max()
+        .unwrap_or(1)
     }
 
     /// Compute the number of blinding factors necessary to perfectly blind
@@ -2234,6 +2259,10 @@ impl<F: Field> ConstraintSystem<F> {
         // - Each lookup argument has independent witness polynomials, and they are
         //   evaluated at most 2 times.
         let factors = std::cmp::max(3, factors);
+
+        // Each trash argument gives away 1 evaluation of a linear combination of some
+        // witness polynomials (the evaluation of the trash column).
+        let factors = factors + self.trashcans.len();
 
         // Each polynomial is evaluated at most an additional time during
         // multiopen (at x_3 to produce q_evals):
@@ -2337,6 +2366,11 @@ impl<F: Field> ConstraintSystem<F> {
         &self.lookups
     }
 
+    /// Returns trash arguments
+    pub fn trashcans(&self) -> &Vec<trash::Argument<F>> {
+        &self.trashcans
+    }
+
     /// Returns constants
     pub fn constants(&self) -> &Vec<Column<Fixed>> {
         &self.constants
@@ -2424,6 +2458,13 @@ impl<'a, F: Field> VirtualCells<'a, F> {
                         poly: q.clone() * constraint.poly,
                     })
                     .collect()
+            }
+            SelectorType::Additive(s) => {
+                let q = self.query_selector(s);
+                let names: Vec<_> = c.constraints.iter().map(|c| c.name.clone()).collect();
+                let polys: Vec<_> = c.constraints.into_iter().map(|c| c.poly).collect();
+                (self.meta.trashcans).push(Argument::new(names.join("&"), q, polys));
+                vec![]
             }
             SelectorType::None => c.constraints,
         }